+++ RELEASE-NOTES.md
@@ -34,6 +34,7 @@ as release notes for end user on a Recalbox upgrade.
 - Fix resolution selection in frontend for Raspberry Pi 4/400/3/Zero2 (KMS)
 - Add new system 'Z-Machine' with new core frotz
 - Add new system 'Macintosh' with new core libretro-MinivMac (RPi3/RPi4/XU4/OGS/PC)
+- Add support for nVidia Prime Offloading (for laptop with both iGPU and nVidia GPU)
 
 ### Improvements
 - Bump libretro pcsx-rearmed